+++ app.js
@@ -1,34 +1,16 @@
-const { promisify } = require("util");
-const { App, ExpressReceiver } = require("@slack/bolt");
-const daraz = require("./");
+const { App: BoltApp } = require("@slack/bolt");
 
-class Receiver extends ExpressReceiver {
+class App extends BoltApp {
   /**
-   * @param {import("express").Request} req
-   * @param {import("express").Response} res
+   * @param  {import("@slack/bolt").AppOptions} opt
    */
-  requestHandler(req, res) {
-    // NOTE: See also https://api.slack.com/events-api#errors
-    res.header("x-slack-no-retry", "1");
-    if (req.headers["x-slack-retry-reason"] === "http_timeout")
-      return promisify(res.end)();
-
-    return super.requestHandler(req, res);
+  constructor(opt) {
+    super(opt);
+    const scripts = require("glob").sync("./scripts/*.js").map(require);
+    scripts.forEach((script) => {
+      if (Array.isArray(script)) return this.message(...script);
+      if (script instanceof Function) return script(this);
+    });
   }
 }
-
-const receiver = new Receiver({
-  signingSecret: process.env.SLACK_SIGNING_SECRET,
-});
-
-const app = daraz(
-  new App({
-    token: process.env.SLACK_BOT_TOKEN,
-    receiver,
-  })
-);
-
-(async () => {
-  await app.start(process.env.PORT || 8080);
-  console.log("Daraz-san ⚡ running");
-})();
+module.exports = App;
