+++ package.json
@@ -2,7 +2,6 @@
   "name": "daraz-san",
   "version": "1.0.0",
   "description": "Daraz-san ⚡",
-  "main": "index.js",
   "author": "Kohei Watanabe <kou029w@gmail.com>",
   "contributors": [
     "SAKAGUCHI Takashi <takashi.sakaguchi@ummm.info>",
@@ -12,7 +11,7 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "start": "node app.js"
+    "start": "node main.js"
   },
   "dependencies": {
     "@slack/bolt": "2.x",
